# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SnapFetch Video Downloader API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnalyzeInput {
  /** The video URL to analyze */
  url: string;
}

export interface DownloadFormat {
  /** Quality label e.g. "1080p", "720p", "360p", "Audio Only" */
  quality: string;
  /** File format e.g. "mp4", "mp3", "webm" */
  format: string;
  /** Human-friendly label for this format */
  label: string;
  /** Direct download URL */
  url: string;
  /**
   * Approximate file size e.g. "45 MB"
   * @nullable
   */
  size?: string | null;
}

export interface VideoInfo {
  title: string;
  /** Detected platform e.g. "youtube", "tiktok", "instagram" */
  platform: string;
  /** Thumbnail image URL */
  thumbnail: string;
  /**
   * Video duration e.g. "3:45"
   * @nullable
   */
  duration: string | null;
  /**
   * Content creator name
   * @nullable
   */
  author?: string | null;
  formats: DownloadFormat[];
}

export interface PlatformStat {
  platform: string;
  count: number;
  percentage: number;
}

export interface FormatStat {
  format: string;
  count: number;
}

export interface DownloadStats {
  totalDownloads: number;
  platformBreakdown: PlatformStat[];
  popularFormats: FormatStat[];
}

export interface BlogPost {
  id: number;
  slug: string;
  title: string;
  excerpt: string;
  content: string;
  category: string;
  publishedAt: string;
  readTime: string;
  /** @nullable */
  coverImage?: string | null;
  author?: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListBlogPostsParams = {
  category?: string;
  limit?: number;
  offset?: number;
};