# lib/api-zod/src/generated/types/analyzeInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SnapFetch Video Downloader API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeInput {
  /** The video URL to analyze */
  url: string;
}