# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SnapFetch Video Downloader API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: number;
  slug: string;
  title: string;
  excerpt: string;
  content: string;
  category: string;
  publishedAt: string;
  readTime: string;
  /** @nullable */
  coverImage?: string | null;
  author?: string;
}