# lib/api-zod/src/generated/types/downloadFormat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SnapFetch Video Downloader API
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadFormat {
  /** Quality label e.g. "1080p", "720p", "360p", "Audio Only" */
  quality: string;
  /** File format e.g. "mp4", "mp3", "webm" */
  format: string;
  /** Human-friendly label for this format */
  label: string;
  /** Direct download URL */
  url: string;
  /**
   * Approximate file size e.g. "45 MB"
   * @nullable
   */
  size?: string | null;
}