# lib/api-zod/src/generated/types/downloadStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SnapFetch Video Downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { FormatStat } from "./formatStat";
import type { PlatformStat } from "./platformStat";

export interface DownloadStats {
  totalDownloads: number;
  platformBreakdown: PlatformStat[];
  popularFormats: FormatStat[];
}