# lib/api-zod/src/generated/types/videoInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SnapFetch Video Downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { DownloadFormat } from "./downloadFormat";

export interface VideoInfo {
  title: string;
  /** Detected platform e.g. "youtube", "tiktok", "instagram" */
  platform: string;
  /** Thumbnail image URL */
  thumbnail: string;
  /**
   * Video duration e.g. "3:45"
   * @nullable
   */
  duration: string | null;
  /**
   * Content creator name
   * @nullable
   */
  author?: string | null;
  formats: DownloadFormat[];
}